Chore: update and run linters
This follows the guidance from https://storybook.js.org/blog/how-to-use-storybook-with-eslint/ to get eslinter set up in the most helpful way with storybook.

I've also run both, which now make modest and helpful changes.

The CSS linter still flags 644 warnings, which we either need to resolve or update the linter configure.