feat: remove all jQuery-dependent legacy components
Here we retire jQuery-specific code from components that we do not use (and are likely to retire or rewrite as they were part of the initial UNDP import).

This allows us to fully remove jQuery from the Mangrove bundle, enhancing performance, reducing the effort to maintain and understand the code base, and allowing us to prepare for future work.